+++ Libplanet.Tests/Blockchain/BlockChainTest.Internals.cs
@@ -7,7 +7,6 @@ using Libplanet.Assets;
 using Libplanet.Blockchain;
 using Libplanet.Blocks;
 using Libplanet.Crypto;
-using Libplanet.Store;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tx;
 using Serilog;
@@ -117,16 +116,9 @@ namespace Libplanet.Tests.Blockchain
             Assert.Equal(expected, actual);
         }
 
-        [Theory]
-        [InlineData(true)]
-        [InlineData(false)]
-        public void ExecuteActions(bool getTxExecutionViaStore)
+        [Fact]
+        public void ExecuteActions()
         {
-            Func<BlockHash, TxId, TxExecution> getTxExecution
-                = getTxExecutionViaStore
-                ? (Func<BlockHash, TxId, TxExecution>)_blockChain.Store.GetTxExecution
-                : _blockChain.GetTxExecution;
-
             (var addresses, Transaction<DumbAction>[] txs) =
                 MakeFixturesForAppendTests();
             var genesis = _blockChain.Genesis;
@@ -144,13 +136,6 @@ namespace Libplanet.Tests.Blockchain
                 renderBlocks: true,
                 renderActions: false
             );
-
-            foreach (Transaction<DumbAction> tx in txs)
-            {
-                Assert.Null(getTxExecution(block1.Hash, tx.Id));
-                Assert.Null(getTxExecution(genesis.Hash, tx.Id));
-                Assert.Null(getTxExecution(_fx.Hash3, tx.Id));
-            }
 
             var minerAddress = genesis.Miner.GetValueOrDefault();
 
@@ -171,143 +156,87 @@ namespace Libplanet.Tests.Blockchain
                     pair.Value,
                     _fx.StateStore.GetState(KeyConverters.ToStateKey(pair.Key), block1.Hash));
             }
+        }
 
-            foreach (Transaction<DumbAction> tx in txs)
+        [Theory]
+        [InlineData(true)]
+        [InlineData(false)]
+        public void UpdateTxExecutions(bool getTxExecutionViaStore)
+        {
+            void AssertTxSuccessesEqual(TxSuccess expected, TxExecution actual)
             {
-                var e = getTxExecution(block1.Hash, tx.Id);
-                Assert.IsType<TxSuccess>(e);
-                var s = (TxSuccess)e;
-                Assert.Equal(block1.Hash, s.BlockHash);
-                Assert.Equal(tx.Id, s.TxId);
-                Assert.Equal(tx.UpdatedAddresses, s.UpdatedAddresses);
-                Assert.Equal(
-                    tx.UpdatedAddresses.ToImmutableDictionary(
-                        address => address,
-                        address => expectedStates[address]
-                    ),
-                    s.UpdatedStates
-                );
-                Assert.Empty(s.FungibleAssetsDelta);
-                Assert.Empty(s.UpdatedFungibleAssets);
-
-                Assert.Null(getTxExecution(genesis.Hash, tx.Id));
-                Assert.Null(getTxExecution(_fx.Hash3, tx.Id));
+                Assert.IsType<TxSuccess>(actual);
+                var success = (TxSuccess)actual;
+                Assert.Equal(expected.TxId, success.TxId);
+                Assert.Equal(expected.BlockHash, success.BlockHash);
+                Assert.Equal(expected.UpdatedStates, success.UpdatedStates);
+                Assert.Equal(expected.FungibleAssetsDelta, success.FungibleAssetsDelta);
+                Assert.Equal(expected.UpdatedFungibleAssets, success.UpdatedFungibleAssets);
             }
 
-            var pk = new PrivateKey();
-            Transaction<DumbAction> tx1Transfer = _fx.MakeTransaction(
-                new[]
-                {
-                    new DumbAction(pk.ToAddress(), "foo", pk.ToAddress(), addresses[1], 10),
-                    new DumbAction(addresses[0], "bar", pk.ToAddress(), addresses[2], 20),
-                },
-                nonce: 0,
-                privateKey: pk
-            );
-            Transaction<DumbAction> tx2Error = _fx.MakeTransaction(
-                new[]
-                {
-                    // As it tries to transfer a negative value, it throws
-                    // ArgumentOutOfRangeException:
-                    new DumbAction(pk.ToAddress(), "foo", addresses[0], addresses[1], -5),
-                },
-                nonce: 1,
-                privateKey: pk
-            );
-            Transaction<DumbAction> tx3Transfer = _fx.MakeTransaction(
-                new[]
-                {
-                    new DumbAction(pk.ToAddress(), "foo", pk.ToAddress(), addresses[1], 5),
-                },
-                nonce: 2,
-                privateKey: pk
-            );
-            Block<DumbAction> block2 = TestUtils.MineNext(
-                genesis,
-                new[] { tx1Transfer, tx2Error, tx3Transfer },
-                difficulty: _blockChain.Policy.GetNextBlockDifficulty(_blockChain)
-            ).AttachStateRootHash(_fx.StateStore, _policy.BlockAction);
-            _blockChain.ExecuteActions(block2);
-            var txExecution1 = getTxExecution(block2.Hash, tx1Transfer.Id);
-            Assert.IsType<TxSuccess>(txExecution1);
-            var txSuccess1 = (TxSuccess)txExecution1;
-            Assert.Equal(
-                addresses.Take(3).Append(pk.ToAddress()).ToImmutableHashSet(),
-                txSuccess1.UpdatedAddresses
-            );
-            Assert.Equal(
-                ImmutableDictionary<Address, IValue>.Empty
-                    .Add(pk.ToAddress(), (Text)"foo")
-                    .Add(addresses[0], (Text)"bar"),
-                txSuccess1.UpdatedStates
-            );
-            Assert.Equal(
+            void AssertTxFailuresEqual(TxFailure expected, TxExecution actual)
+            {
+                Assert.IsType<TxFailure>(actual);
+                var failure = (TxFailure)actual;
+                Assert.Equal(expected.TxId, failure.TxId);
+                Assert.Equal(expected.BlockHash, failure.BlockHash);
+                Assert.Equal(expected.ExceptionName, failure.ExceptionName);
+                Assert.Equal(expected.ExceptionMetadata, failure.ExceptionMetadata);
+            }
+
+            Func<BlockHash, TxId, TxExecution> getTxExecution
+                = getTxExecutionViaStore
+                ? (Func<BlockHash, TxId, TxExecution>)_blockChain.Store.GetTxExecution
+                : _blockChain.GetTxExecution;
+
+            Assert.Null(getTxExecution(_fx.Hash1, _fx.TxId1));
+            Assert.Null(getTxExecution(_fx.Hash1, _fx.TxId2));
+            Assert.Null(getTxExecution(_fx.Hash2, _fx.TxId1));
+            Assert.Null(getTxExecution(_fx.Hash2, _fx.TxId2));
+
+            var random = new System.Random();
+            var inputA = new TxSuccess(
+                _fx.Hash1,
+                _fx.TxId1,
+                ImmutableDictionary<Address, IValue>.Empty.Add(
+                    random.NextAddress(),
+                    (Text)"state value"
+                ),
                 ImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>>.Empty
                     .Add(
-                        pk.ToAddress(),
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * -30)
-                    )
-                    .Add(
-                        addresses[1],
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * 10)
-                    )
-                    .Add(
-                        addresses[2],
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * 20)
+                        random.NextAddress(),
+                        ImmutableDictionary<Currency, FAV>.Empty.Add(
+                            DumbAction.DumbCurrency,
+                            DumbAction.DumbCurrency * 5
+                        )
                     ),
-                txSuccess1.UpdatedFungibleAssets
-            );
-            Assert.Equal(
-                txSuccess1.FungibleAssetsDelta,
-                txSuccess1.UpdatedFungibleAssets
-            );
-            var txExecution2 = getTxExecution(block2.Hash, tx2Error.Id);
-            Assert.IsType<TxFailure>(txExecution2);
-            var txFailure = (TxFailure)txExecution2;
-            Assert.Equal(block2.Hash, txFailure.BlockHash);
-            Assert.Equal(tx2Error.Id, txFailure.TxId);
-            Assert.Equal(
-                $"{nameof(System)}.{nameof(ArgumentOutOfRangeException)}",
-                txFailure.ExceptionName
-            );
-            Assert.Equal(
-                Dictionary.Empty.Add("parameterName", "value"),
-                txFailure.ExceptionMetadata
-            );
-            var txExecution3 = getTxExecution(block2.Hash, tx3Transfer.Id);
-            Assert.IsType<TxSuccess>(txExecution3);
-            var txSuccess3 = (TxSuccess)txExecution3;
-            Assert.Equal(
                 ImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>>.Empty
                     .Add(
-                        pk.ToAddress(),
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * -5)
+                        random.NextAddress(),
+                        ImmutableDictionary<Currency, FAV>.Empty.Add(
+                            DumbAction.DumbCurrency,
+                            DumbAction.DumbCurrency * 10
+                        )
                     )
-                    .Add(
-                        addresses[1],
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * 5)
-                    ),
-                txSuccess3.FungibleAssetsDelta
             );
-            Assert.Equal(
-                ImmutableDictionary<Address, IImmutableDictionary<Currency, FAV>>.Empty
-                    .Add(
-                        pk.ToAddress(),
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * -35)
-                    )
-                    .Add(
-                        addresses[1],
-                        ImmutableDictionary<Currency, FAV>.Empty
-                            .Add(DumbAction.DumbCurrency, DumbAction.DumbCurrency * 15)
-                    ),
-                txSuccess3.UpdatedFungibleAssets
+            var inputB = new TxFailure(
+                _fx.Hash1,
+                _fx.TxId2,
+                "AnExceptionName",
+                Dictionary.Empty.Add("foo", 1).Add("bar", "baz")
             );
+            var inputC = new TxFailure(
+                _fx.Hash2,
+                _fx.TxId1,
+                "AnotherExceptionName",
+                null
+            );
+            _blockChain.UpdateTxExecutions(new TxExecution[] { inputA, inputB, inputC });
+
+            AssertTxSuccessesEqual(inputA, getTxExecution(_fx.Hash1, _fx.TxId1));
+            AssertTxFailuresEqual(inputB, getTxExecution(_fx.Hash1, _fx.TxId2));
+            AssertTxFailuresEqual(inputC, getTxExecution(_fx.Hash2, _fx.TxId1));
+            Assert.Null(getTxExecution(_fx.Hash2, _fx.TxId2));
         }
     }
 }