+++ Libplanet/Action/ActionContext.cs
@@ -1,7 +1,6 @@
 #nullable enable
 using System.Diagnostics.Contracts;
 using System.Security.Cryptography;
-using Libplanet.Blocks;
 using Libplanet.Store.Trie;
 using Libplanet.Tx;
 
@@ -18,7 +17,6 @@ namespace Libplanet.Action
             Address signer,
             TxId? txid,
             Address miner,
-            BlockHash blockHash,
             long blockIndex,
             IAccountStateDelta previousStates,
             int randomSeed,
@@ -30,7 +28,6 @@ namespace Libplanet.Action
             Signer = signer;
             TxId = txid;
             Miner = miner;
-            BlockHash = blockHash;
             BlockIndex = blockIndex;
             Rehearsal = rehearsal;
             PreviousStates = previousStates;
@@ -45,8 +42,6 @@ namespace Libplanet.Action
         public TxId? TxId { get; }
 
         public Address Miner { get; }
-
-        public BlockHash BlockHash { get; }
 
         public long BlockIndex { get; }
 
@@ -75,7 +70,6 @@ namespace Libplanet.Action
                 Signer,
                 TxId,
                 Miner,
-                BlockHash,
                 BlockIndex,
                 PreviousStates,
                 _randomSeed,